requirements.txt
streamlit
fastapi
uvicorn
pymongo
pdfplumber
spacy
python-dotenv
google-generativeai
python-multipart
